move split-related tests
