+++ tests/test_splitting.py
@@ -1,4 +1,6 @@
 """Tests for splitting of triples."""
+import numpy
+import pytest
 import torch
 
 from pykeen.triples.splitting import (
@@ -7,10 +9,71 @@ from pykeen.triples.splitting import (
     DeterministicCleaner,
     RandomizedCleaner,
     _get_cover_deterministic,
+    get_absolute_split_sizes,
+    normalize_ratios,
 )
 from pykeen.triples.utils import get_entities, get_relations
 from pykeen.utils import tensor_to_set
 from tests.cases import CleanerTestCase, SplitterTestCase
+
+
+def test_get_absolute_split_sizes():
+    """Test get_absolute_split_sizes."""
+    for num_splits, n_total in zip(
+        (2, 3, 4),
+        (100, 200, 10412),
+    ):
+        # generate random ratios
+        ratios = numpy.random.uniform(size=(num_splits,))
+        ratios = ratios / ratios.sum()
+        sizes = get_absolute_split_sizes(n_total=n_total, ratios=ratios)
+        # check size
+        assert len(sizes) == len(ratios)
+
+        # check value range
+        assert all(0 <= size <= n_total for size in sizes)
+
+        # check total split
+        assert sum(sizes) == n_total
+
+        # check consistency with ratios
+        rel_size = numpy.asarray(sizes) / n_total
+        # the number of decimal digits equivalent to 1 / n_total
+        decimal = numpy.floor(numpy.log10(n_total))
+        numpy.testing.assert_almost_equal(rel_size, ratios, decimal=decimal)
+
+
+def test_normalize_ratios():
+    """Test normalize_ratios."""
+    for ratios, exp_output in (
+        (0.5, (0.5, 0.5)),
+        ((0.3, 0.2, 0.4), (0.3, 0.2, 0.4, 0.1)),
+        ((0.3, 0.3, 0.4), (0.3, 0.3, 0.4)),
+    ):
+        output = normalize_ratios(ratios=ratios)
+        # check type
+        assert isinstance(output, tuple)
+        assert all(isinstance(ratio, float) for ratio in output)
+        # check values
+        assert len(output) >= 2
+        assert all(0 <= ratio <= 1 for ratio in output)
+        output_np = numpy.asarray(output)
+        numpy.testing.assert_almost_equal(output_np.sum(), numpy.ones(1))
+        # compare against expected
+        numpy.testing.assert_almost_equal(output_np, numpy.asarray(exp_output))
+
+
+def test_normalize_invalid_ratio():
+    """Test invalid ratios."""
+    cases = [
+        1.1,
+        [1.1],
+        [0.8, 0.3],
+        [0.8, 0.1, 0.2],
+    ]
+    for ratios in cases:
+        with pytest.raises(ValueError):
+            _ = normalize_ratios(ratios=ratios)
 
 
 class DeterministicCleanerTests(CleanerTestCase):
